requirements.txt
crewai
crewai_tools
duckduckgo-search
python-dotenv
langchain-openai